SRVインデックス割り当てのコメントアウトと型キャスト修正
`TextureManager::LoadTexture` 関数内でSRVインデックスの割り当てコードをコメントアウトしました。具体的には、`uint32_t srvIndex = srvSetup_->Allocate();` の行がコメントアウトされました。

`SrvSetup::CreateSRVStructuredBuffer` 関数内で、`NumElements` の計算において、`pResource->GetDesc().Width / structureByteStride` の結果を `UINT` 型にキャストするように変更しました。

+++ engine/base/SrvSetup.cpp
@@ -75,7 +75,7 @@ void SrvSetup::CreateSRVStructuredBuffer(uint32_t srvIndex, ID3D12Resource* pRes
 	srvDesc.Shader4ComponentMapping = D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING;
 	srvDesc.Buffer.StructureByteStride = structureByteStride;
 	srvDesc.Buffer.FirstElement = 0;
-	srvDesc.Buffer.NumElements = pResource->GetDesc().Width / structureByteStride;
+	srvDesc.Buffer.NumElements = static_cast<UINT>(pResource->GetDesc().Width / structureByteStride);
 	srvDesc.Buffer.Flags = D3D12_BUFFER_SRV_FLAG_NONE;
 	dxCore_->GetDevice()->CreateShaderResourceView(pResource, &srvDesc, handleCPU);
 }
